refactor spdata module to sdata module
