+++ src/scportrait/pipeline/project.py
@@ -34,7 +34,7 @@ from scportrait.pipeline._utils.spatialdata_helper import (
     get_chunk_size,
     rechunk_image,
 )
-from scportrait.tools.spdata.write._helper import _get_image, _get_shape
+from scportrait.tools.sdata.write._helper import _get_image, _get_shape
 
 if TYPE_CHECKING:
     from collections.abc import Callable
