+++ src/scportrait/tools/__init__.py
@@ -1,3 +1,3 @@
-from . import h5sc, ml, parse, spdata, stitch
+from . import h5sc, ml, parse, sdata, stitch
 
-__all__ = ["h5sc", "ml", "parse", "stitch", "spdata"]
+__all__ = ["h5sc", "ml", "parse", "stitch", "sdata"]
