+++ src/scportrait/tools/sdata/processing/__init__.py
@@ -1,3 +1,4 @@
 from ._image_processing import percentile_normalize_image
+from ._subset import get_bounding_box_sdata
 
 __all__ = ["percentile_normalize_image"]
