+++ src/scportrait/tools/sdata/write/_helper.py
